@@ -14,7 +14,7 @@
         <artifactId>maven-compiler-plugin</artifactId>
         <version>3.10.0</version>
         <configuration>
-          <release>17</release>
+          <release>23</release>
         </configuration>
       </plugin>
     </plugins>
