@@ -19,12 +19,14 @@ implements PrettyPrintable {
 	@Override
 	public void mkString(PrettyPrinter pp) {
 		pp.append(pattern).append(" ->").inc().endl();
-		pp.append(body).dec();
+		pp.indent(() -> {
+			pp.append(body);
+		});
 	}
 
 	CcCaseBranch substId(IdMap<Id> map) {
 		Set<Id> ids = new HashSet<>();
-		pattern.addVars(ids);
+		pattern.accumulateVars(ids);
 		ids.forEach(id -> {
 			if(map.containsKey(id)) {
 				throw new RuntimeException(""+id);
