@@ -8,7 +8,7 @@ import zlk.util.LocationHolder;
 import zlk.util.pp.PrettyPrintable;
 import zlk.util.pp.PrettyPrinter;
 
-public record CcType(
+public record CcTypeDecl(
 		Id id,
 		List<CcCtor> ctors,
 		Location loc
@@ -16,11 +16,12 @@ public record CcType(
 
 		@Override
 		public void mkString(PrettyPrinter pp) {
-			pp.append("type ").append(id).endl().inc();
-			pp.append("= ").append(ctors.get(0));
-			ctors.subList(1, ctors.size())
-					.forEach(ctor -> pp.endl().append("| ").append(ctor));
-			pp.dec();
+			pp.append("type ").append(id).endl();
+			pp.indent(() -> {
+				pp.append("= ").append(ctors.get(0));
+				ctors.subList(1, ctors.size())
+						.forEach(ctor -> pp.endl().append("| ").append(ctor));
+			});
 		}
 
 		@Override
