@@ -10,32 +10,43 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Function;
-import java.util.stream.Stream;
 
-import zlk.clcalc.CcApp;
-import zlk.clcalc.CcCase;
 import zlk.clcalc.CcCaseBranch;
-import zlk.clcalc.CcCnst;
 import zlk.clcalc.CcCtor;
-import zlk.clcalc.CcDecl;
 import zlk.clcalc.CcExp;
-import zlk.clcalc.CcIf;
-import zlk.clcalc.CcLet;
-import zlk.clcalc.CcMkCls;
+import zlk.clcalc.CcExp.CcApp;
+import zlk.clcalc.CcExp.CcCase;
+import zlk.clcalc.CcExp.CcCnst;
+import zlk.clcalc.CcExp.CcIf;
+import zlk.clcalc.CcExp.CcLet;
+import zlk.clcalc.CcExp.CcMkCls;
+import zlk.clcalc.CcExp.CcVar;
+import zlk.clcalc.CcFunDecl;
 import zlk.clcalc.CcModule;
-import zlk.clcalc.CcType;
-import zlk.clcalc.CcVar;
+import zlk.clcalc.CcTypeDecl;
+import zlk.common.ConstValue;
+import zlk.common.Type;
 import zlk.common.id.Id;
 import zlk.common.id.IdList;
 import zlk.common.id.IdMap;
-import zlk.common.type.Type;
+import zlk.idcalc.IcCaseBranch;
 import zlk.idcalc.IcCtor;
-import zlk.idcalc.IcDecl;
 import zlk.idcalc.IcExp;
+import zlk.idcalc.IcExp.IcAbs;
+import zlk.idcalc.IcExp.IcApp;
+import zlk.idcalc.IcExp.IcCase;
+import zlk.idcalc.IcExp.IcCnst;
+import zlk.idcalc.IcExp.IcIf;
+import zlk.idcalc.IcExp.IcLet;
+import zlk.idcalc.IcExp.IcLetrec;
+import zlk.idcalc.IcExp.IcVarCtor;
+import zlk.idcalc.IcExp.IcVarForeign;
+import zlk.idcalc.IcExp.IcVarLocal;
+import zlk.idcalc.IcFunDecl;
 import zlk.idcalc.IcModule;
-import zlk.idcalc.IcPVar;
 import zlk.idcalc.IcPattern;
-import zlk.idcalc.IcType;
+import zlk.idcalc.IcPattern.Var;
+import zlk.idcalc.IcTypeDecl;
 import zlk.util.Location;
 
 public final class ClosureConveter {
@@ -44,7 +55,7 @@ public final class ClosureConveter {
 	private final IdMap<Type> type; // 変換後のIdの型を追加
 	private final Set<Id> knowns;
 
-	private final List<CcDecl> toplevels;
+	private final List<CcFunDecl> toplevels;
 	private final AtomicInteger closureCount;
 
 	public ClosureConveter(IcModule src, IdMap<Type> type, IdList builtins) {
@@ -63,9 +74,9 @@ public final class ClosureConveter {
 				.stream()
 				.map(decl -> compileFunc(decl.id(), decl.args(), decl.body()))
 				.forEach(maybeCls -> maybeCls.ifPresent(cls -> {
-					throw new RuntimeException("toplevel must not be closure: "+cls.id()); }));
+					throw new RuntimeException("toplevel must not be closure: "+cls.clsFunc()); }));
 
-		List<CcType> types = src.types().stream().map(ty -> convert(ty)).toList();
+		List<CcTypeDecl> types = src.types().stream().map(ty -> convert(ty)).toList();
 		return new CcModule(src.name(), types, toplevels, src.origin());
 	}
 
@@ -82,12 +93,12 @@ public final class ClosureConveter {
 
 		if(frees.isEmpty()) {
 			System.out.println(id + " is not cloeure.");
-			toplevels.add(new CcDecl(id, args_, ccBody, body.loc()));
+			toplevels.add(new CcFunDecl(id, args_, ccBody, body.loc()));
 			return Optional.empty();
 
 		} else {
 			System.out.println(id + " is cloeure. frees: "+frees);
-			CcDecl closureFunc = makeClosure(id, frees, args_, ccBody, type.get(id).apply(args_.size()), body.loc());
+			CcFunDecl closureFunc = makeClosure(id, frees, args_, ccBody, type.get(id).apply(args_.size()), body.loc());
 			toplevels.add(closureFunc);
 			knowns.add(closureFunc.id());
 			return Optional.of(new CcMkCls(closureFunc.id(), frees, closureFunc.loc()));
@@ -104,16 +115,12 @@ public final class ClosureConveter {
 	 * @param loc 元の関数のLocation
 	 * @return
 	 */
-	private CcDecl makeClosure(Id original, IdList frees, List<IcPattern> args_, CcExp body, Type retTy, Location loc) {
-		List<Type> types =
-				Stream.concat(Stream.concat(
-						frees.stream().map(type::get),
-						args_.stream()
-								.map(pat -> pat.fold(
-										var -> type.get(var.id()),
-										ctor -> type.get(ctor.ctor().id())))),
-						Stream.of(retTy))
-						.toList();
+	private CcFunDecl makeClosure(Id original, IdList frees, List<IcPattern> args_, CcExp body, Type retTy, Location loc) {
+		List<Type> types = new ArrayList<>();
+		frees.forEach(id -> types.add(type.get(id)));
+		args_.forEach(pat -> types.add(type.get(pat.headId())));
+		types.add(retTy);
+
 		Type clsTy = types.get(0).toTree(types.subList(1, types.size()));
 		Id clsId = freshId(original);
 		type.put(clsId, clsTy);
@@ -125,129 +132,138 @@ public final class ClosureConveter {
 		List<IcPattern> clsArgs = new ArrayList<>(args_);
 		for(Id free : frees) {
 			Id newId = idMap.get(free);
-			clsArgs.add(new IcPVar(newId, Location.noLocation()));
+			clsArgs.add(new Var(newId, Location.noLocation()));
 			type.put(newId, type.get(free));
 		}
 
 		CcExp clsBody = body.substId(idMap);
 
-		return new CcDecl(clsId, clsArgs, clsBody, loc);
+		return new CcFunDecl(clsId, clsArgs, clsBody, loc);
 	}
 
-	private CcExp compile(IcExp body) {
-		return body.fold(
-				cnst    -> new CcCnst(cnst.value(), cnst.loc()),
-				var     -> new CcVar(var.id(), var.loc()),
-				foreign -> new CcVar(foreign.id(), foreign.loc()),
-				ctor    -> new CcVar(ctor.id(), ctor.loc()),
-				abs     -> {
-					return neverHappen(
-							"no anonymous abs in this version.", body.loc());
-				},
-				app     -> {
-					CcExp funExp = compile(app.fun());
-					List<CcExp> argExps =
-							app.args().stream()
-							.map(arg -> compile(arg)).toList();
-					return new CcApp(funExp, argExps, app.loc());
-				},
-				if_     -> new CcIf(
-						compile(if_.cond()),
-						compile(if_.exp1()),
-						compile(if_.exp2()),
-						if_.loc()),
-				let     -> {
-					Id id = let.decl().id();
-					List<IcPattern> args = let.decl().args();
-					IcExp bounded = let.decl().body();
-					CcExp letBody = compile(let.body());
-					Location loc = let.loc();
+	private CcExp compile(IcExp exp) {
+		return switch (exp) {
+		case IcCnst(ConstValue value, Location loc) -> {
+			yield new CcCnst(value, loc);
+		}
+		case IcVarLocal(Id id, Location loc) -> {
+			yield new CcVar(id, loc);
+		}
+		case IcVarForeign(Id id, Type _, Location loc) -> {
+			yield new CcVar(id, loc);
+		}
+		case IcVarCtor(Id id, Type _, Location loc) -> {
+			yield new CcVar(id, loc);
+		}
+		case IcAbs(Id _, Type _, IcExp body, Location _) -> {
+			yield neverHappen("no anonymous abs in this version.", body.loc());
+		}
+		case IcApp(IcExp fun, List<IcExp> args, Location loc) -> {
+			CcExp funExp = compile(fun);
+			List<CcExp> argExps = args.stream()
+					.map(arg -> compile(arg))
+					.toList();
+			yield new CcApp(funExp, argExps, loc);
+		}
+		case IcIf(IcExp cond, IcExp thenExp, IcExp elseExp, Location loc) -> {
+			yield new CcIf(
+					compile(cond),
+					compile(thenExp),
+					compile(elseExp),
+					loc);
+		}
+		case IcLet(IcFunDecl decl, IcExp body, Location loc) -> {
+			Id id = decl.id();
+			List<IcPattern> args = decl.args();
+			IcExp bounded = decl.body();
+			CcExp letBody = compile(body);
 
-					if(!args.isEmpty()) {
-						return compileFunc(id, let.decl().args(), bounded)
-								.map(mkCls -> (CcExp)new CcLet(id, mkCls, letBody, loc))
-								.orElse(letBody); // トップレベルで定義されているのでletは要らない
-					} else {
-						return new CcLet(id, compile(bounded), letBody, loc);
-					}
-				},
-				letrec -> {
-					if(letrec.decls().size() != 1) {
-						todo();
-					}
-					IcDecl decl = letrec.decls().get(0);
-					Id id = decl.id();
-					List<IcPattern> args = decl.args();
-					IcExp bounded = decl.body();
-					CcExp letBody = compile(letrec.body());
-					Location loc = letrec.loc();
-					if(!args.isEmpty()) {
-						return compileFunc(id, decl.args(), bounded)
-								.map(mkCls -> (CcExp)new CcLet(id, mkCls, letBody, loc))
-								.orElse(letBody); // トップレベルで定義されているのでletは要らない
-					} else {
-						return new CcLet(id, compile(bounded), letBody, loc);
-					}
-				},
-				case_ -> {
-					CcExp target = compile(case_.target());
-					List<CcCaseBranch> branches =
-							case_.branches().stream()
-									.map(branch -> new CcCaseBranch(
-											branch.pattern(),
-											compile(branch.body()),
-											branch.loc()))
-									.toList();
-					return new CcCase(target, branches, case_.loc());
-				});
+			if(!args.isEmpty()) {
+				yield compileFunc(id, decl.args(), bounded)
+						.map(mkCls -> (CcExp)new CcLet(id, mkCls, letBody, loc))
+						.orElse(letBody); // トップレベルで定義されているのでletは要らない
+			} else {
+				yield new CcLet(id, compile(bounded), letBody, loc);
+			}
+		}
+		case IcLetrec(List<IcFunDecl> decls, IcExp body, Location loc) -> {
+			if(decls.size() != 1) {
+				todo();
+			}
+			IcFunDecl decl = decls.get(0);
+			Id id = decl.id();
+			List<IcPattern> args = decl.args();
+			IcExp bounded = decl.body();
+			CcExp letBody = compile(body);
+
+			if(!args.isEmpty()) {
+				yield compileFunc(id, decl.args(), bounded)
+						.map(mkCls -> (CcExp)new CcLet(id, mkCls, letBody, loc))
+						.orElse(letBody);
+			} else {
+				yield new CcLet(id, compile(bounded), letBody, loc);
+			}
+		}
+		case IcCase(IcExp target, List<IcCaseBranch> branches, Location loc) -> {
+			CcExp compiledTarget = compile(target);
+			List<CcCaseBranch> compiledBranches =
+					branches.stream()
+							.map(branch -> new CcCaseBranch(
+									branch.pattern(),
+									compile(branch.body()),
+									branch.loc()))
+							.toList();
+			yield new CcCase(compiledTarget, compiledBranches, loc);
+		}
+		};
 	}
 
 	private IdList fvFunc(CcExp body, List<IcPattern> args) {
 		IdList free = new IdList();
 		Set<Id> bounded = new HashSet<>(knowns);
-		args.forEach(arg -> arg.addVars(bounded));
+		args.forEach(arg -> arg.accumulateVars(bounded));
 		fv(body, bounded, free);
 		return free;
 	}
 
 	private void fv(CcExp exp, Set<Id> bounded, IdList free) {
-		exp.match(
-				cnst -> {},
-				var  -> {
-					Id id = var.id();
-					if(!bounded.contains(id) && !free.contains(id)) {
-						free.add(id);
-					}
-				},
-				call  -> {
-					fv(call.fun(), bounded, free);
-					call.args().forEach(arg -> fv(arg, bounded, free));
-				},
-				mkCls -> {
-					if(!bounded.contains(mkCls.clsFunc())) {
-						throw new AssertionError();
-					}
-					mkCls.caps().stream()
-							.filter(id -> !bounded.contains(id) && !free.contains(id))
-							.forEach(free::add);
-				},
-				if_  -> {
-					fv(if_.cond(), bounded, free);
-					fv(if_.thenExp(), bounded, free);
-					fv(if_.elseExp(), bounded, free);
-				},
-				let  -> {
-					bounded.add(let.boundVar());
-					fv(let.boundExp(), bounded, free);
-					fv(let.mainExp(), bounded, free);
-				},
-				case_ -> {
-					fv(case_.target(), bounded, free);
-					for(CcCaseBranch branch: case_.branches()) {
-						branch.pattern().addVars(bounded);
-						fv(branch.body(), bounded, free);
-					}
-				});
+		switch (exp) {
+		case CcCnst _ -> {
+		}
+		case CcVar(Id id, Location _) -> {
+			if (!bounded.contains(id) && !free.contains(id)) {
+				free.add(id);
+			}
+		}
+		case CcApp(CcExp fun, List<CcExp> args, Location _) -> {
+			fv(fun, bounded, free);
+			args.forEach(arg -> fv(arg, bounded, free));
+		}
+		case CcMkCls(Id clsFunc, IdList caps, Location _) -> {
+			if (!bounded.contains(clsFunc)) {
+				throw new AssertionError();
+			}
+			caps.stream().filter(id -> !bounded.contains(id) && !free.contains(id))
+					.forEach(free::add);
+		}
+		case CcIf(CcExp cond, CcExp thenExp, CcExp elseExp, Location _) -> {
+			fv(cond, bounded, free);
+			fv(thenExp, bounded, free);
+			fv(elseExp, bounded, free);
+		}
+		case CcLet(Id varName, CcExp boundExp, CcExp body, Location _) -> {
+			bounded.add(varName);
+			fv(boundExp, bounded, free);
+			fv(body, bounded, free);
+		}
+		case CcCase(CcExp target, List<CcCaseBranch> branches, Location _) -> {
+			fv(target, bounded, free);
+			for (CcCaseBranch branch : branches) {
+				branch.pattern().accumulateVars(bounded);
+				fv(branch.body(), bounded, free);
+			}
+		}
+		}
 	}
 
 	private Id freshId(Id original) {
@@ -265,8 +281,8 @@ public final class ClosureConveter {
 				+ Id.SEPARATOR + exceptModule);
 	}
 
-	private CcType convert(IcType icType) {
-		return new CcType(icType.id(), icType.ctors().stream().map(ctor -> convert(ctor)).toList(), icType.loc());
+	private CcTypeDecl convert(IcTypeDecl icType) {
+		return new CcTypeDecl(icType.id(), icType.ctors().stream().map(ctor -> convert(ctor)).toList(), icType.loc());
 	}
 
 	private CcCtor convert(IcCtor icCtor) {
