@@ -1,20 +1,21 @@
 package zlk.idcalc;
 
+import java.util.Iterator;
 import java.util.List;
 import java.util.Optional;
 
+import zlk.common.Type;
 import zlk.common.id.Id;
-import zlk.common.type.Type;
 import zlk.util.Location;
 import zlk.util.LocationHolder;
 import zlk.util.pp.PrettyPrintable;
 import zlk.util.pp.PrettyPrinter;
 
-public record IcDecl(
+public record IcFunDecl(
 		Id id,
 		Optional<Type> anno,
 		List<IcPattern> args,
-		Optional<List<IcDecl>> recs,
+		Optional<List<IcFunDecl>> recs,
 		IcExp body,
 		Location loc)
 implements PrettyPrintable, LocationHolder {
@@ -25,14 +26,16 @@ implements PrettyPrintable, LocationHolder {
 
 	@Override
 	public void mkString(PrettyPrinter pp) {
-		pp.append(id);
-		if(recs.isPresent()) {
-			pp.append(" [");
-			pp.oneline(recs.get().stream().map(decl -> decl.id).toList(), ", ");
-			pp.append("]");
-		}
+		pp.append(id).append(" ");
+		recs.ifPresent(recList -> {
+			Iterator<Id> idIter = recList.stream().map(decl -> decl.id).iterator();
+			PrettyPrintable elmStyleIdList = PrettyPrintable.toElmListStyle(idIter);
+			pp.withoutLineBreak(
+					pp_ -> pp_.append(elmStyleIdList)
+			);
+		});
 		anno.ifPresent(ty -> {
-			pp.append(" : ").append(ty).endl();
+			pp.append(": ").append(ty).endl();
 			pp.append(id);
 		});
 		args.forEach(arg -> {
@@ -42,7 +45,7 @@ implements PrettyPrintable, LocationHolder {
 		pp.inc().append(body).dec();
 	}
 
-	public IcDecl norec() {
-		return new IcDecl(id, anno, args, Optional.empty(), body, loc);
+	public IcFunDecl norec() {
+		return new IcFunDecl(id, anno, args, Optional.empty(), body, loc);
 	}
 }
